added routes js to route api calls through index and added separate modules for music and tube
